refactor(session): split projection ownership
