+++ packages/projection/src/pi-event-display.ts
@@ -107,117 +107,6 @@ const oneLine = (value: string) =>
 		.trim();
 const truncate = (value: string, max = 96) =>
 	value.length <= max ? value : `${value.slice(0, Math.max(0, max - 1))}…`;
-const compactStringField = (key: string, value: unknown) =>
-	typeof value === "string" ? { [key]: value } : {};
-const compactBooleanField = (key: string, value: unknown) =>
-	typeof value === "boolean" ? { [key]: value } : {};
-const compactNumberField = (key: string, value: unknown) =>
-	typeof value === "number" ? { [key]: value } : {};
-
-const compactToolArgs = (value: unknown): Record<string, unknown> => {
-	const args = isRecord(value) ? value : {};
-	return {
-		...compactStringField("command", args["command"]),
-		...compactStringField("path", args["path"]),
-		...compactStringField("pattern", args["pattern"]),
-	};
-};
-
-const compactUsage = (value: unknown): Record<string, unknown> | undefined => {
-	const usage = isRecord(value) ? value : undefined;
-	if (usage === undefined) return undefined;
-	return {
-		...compactNumberField("input", usage["input"]),
-		...compactNumberField("inputTokens", usage["inputTokens"]),
-		...compactNumberField("output", usage["output"]),
-		...compactNumberField("outputTokens", usage["outputTokens"]),
-		...compactNumberField("total", usage["total"]),
-		...compactNumberField("totalTokens", usage["totalTokens"]),
-		...(isRecord(usage["cost"])
-			? { cost: compactNumberField("total", usage["cost"]["total"]) }
-			: {}),
-	};
-};
-
-const compactContentItem = (value: unknown): unknown => {
-	if (!isRecord(value)) return value;
-	const type = str(value["type"]);
-	if (type === "thinking")
-		return { type, ...compactStringField("thinking", value["thinking"]) };
-	if (type === "text")
-		return { type, ...compactStringField("text", value["text"]) };
-	if (type === "toolCall")
-		return {
-			type,
-			...compactStringField("id", value["id"]),
-			...compactStringField("name", value["name"]),
-			arguments: compactToolArgs(value["arguments"]),
-		};
-	return type === undefined ? {} : { type };
-};
-
-const compactPartial = (
-	value: unknown,
-): Record<string, unknown> | undefined => {
-	const partial = isRecord(value) ? value : undefined;
-	if (partial === undefined) return undefined;
-	return Array.isArray(partial["content"])
-		? { content: partial["content"].map(compactContentItem) }
-		: {};
-};
-
-const compactAssistantMessageEvent = (
-	value: unknown,
-): Record<string, unknown> | undefined => {
-	const update = isRecord(value) ? value : undefined;
-	if (update === undefined) return undefined;
-	const partial = compactPartial(update["partial"]);
-	return {
-		...compactStringField("type", update["type"]),
-		...compactStringField("delta", update["delta"]),
-		...compactNumberField("contentIndex", update["contentIndex"]),
-		...(partial === undefined ? {} : { partial }),
-	};
-};
-
-const compactMessage = (
-	value: unknown,
-): Record<string, unknown> | undefined => {
-	const message = isRecord(value) ? value : undefined;
-	if (message === undefined) return undefined;
-	const usage = compactUsage(message["usage"]);
-	return {
-		...compactStringField("responseId", message["responseId"]),
-		...compactNumberField("timestamp", message["timestamp"]),
-		...(usage === undefined ? {} : { usage }),
-	};
-};
-
-export const compactPiEvent = (value: unknown): Record<string, unknown> => {
-	const event = isRecord(value) ? value : {};
-	const usage = compactUsage(event["usage"]);
-	const message = compactMessage(event["message"]);
-	const assistantMessageEvent = compactAssistantMessageEvent(
-		event["assistantMessageEvent"],
-	);
-	return {
-		...compactStringField("type", event["type"]),
-		...compactStringField("sessionFile", event["sessionFile"]),
-		...compactStringField("sessionId", event["sessionId"]),
-		...compactStringField("toolName", event["toolName"]),
-		...compactStringField("toolCallId", event["toolCallId"]),
-		...compactStringField("text", event["text"]),
-		...compactStringField("content", event["content"]),
-		...compactBooleanField("isError", event["isError"]),
-		...(isRecord(event["args"])
-			? { args: compactToolArgs(event["args"]) }
-			: {}),
-		...(usage === undefined ? {} : { usage }),
-		...(message === undefined ? {} : { message }),
-		...(assistantMessageEvent === undefined ? {} : { assistantMessageEvent }),
-	};
-};
-
 const contentText = (value: unknown, kind: "thinking" | "text") => {
 	const partial = isRecord(value) ? value : undefined;
 	const content = Array.isArray(partial?.["content"]) ? partial["content"] : [];
