+++ packages/projection/src/projection-parts/helpers.ts
