+++ packages/projection/src/projection-parts/snapshot.ts
