+++ packages/projection/src/projection-parts/work.ts
