+++ packages/projection/tsconfig.json
@@ -1,0 +1,9 @@
+{
+	"extends": "@repo/tsconfig/base.json",
+	"compilerOptions": {
+		"rootDir": ".",
+		"noEmit": true,
+		"allowImportingTsExtensions": true
+	},
+	"include": ["src/**/*.ts", "test/**/*.ts"]
+}
