+++ packages/session/src/runtime-event.ts
@@ -1,11 +1,6 @@
 import { randomUUID } from "node:crypto";
 import { Schema } from "effect";
-import {
-	NonEmptyString,
-	PositiveInteger,
-	decodeBoundary,
-	optional,
-} from "./schema.js";
+import { NonEmptyString, PositiveInteger, optional } from "./schema.js";
 
 export const sessionEventSchema = Schema.Struct({
 	kind: Schema.Literal("session_event"),
@@ -48,43 +43,42 @@ export interface AgentEventInput {
 	readonly timestamp?: string;
 }
 
-const optionalString = (key: string, value: string | undefined) =>
-	value === undefined ? {} : { [key]: value };
-
 export const makeSessionEventRecord = (input: {
 	readonly sessionId: string;
 	readonly sequence: number;
 	readonly timestamp?: string;
 	readonly event: SessionEventInput;
-}): RuntimeEvent =>
-	decodeBoundary(runtimeEventSchema, {
-		kind: "session_event",
-		sessionId: input.sessionId,
-		sequence: input.sequence,
-		timestamp:
-			input.event.timestamp ?? input.timestamp ?? new Date().toISOString(),
-		type: input.event.type,
-		...(input.event.payload === undefined
-			? {}
-			: { payload: input.event.payload }),
-	});
+}): RuntimeEvent => ({
+	kind: "session_event",
+	sessionId: input.sessionId,
+	sequence: input.sequence,
+	timestamp:
+		input.event.timestamp ?? input.timestamp ?? new Date().toISOString(),
+	type: input.event.type,
+	...(input.event.payload === undefined
+		? {}
+		: { payload: input.event.payload }),
+});
 
 export const makeAgentEventRecord = (input: {
 	readonly sessionId: string;
 	readonly sequence: number;
 	readonly timestamp?: string;
 	readonly event: AgentEventInput;
-}): RuntimeEvent =>
-	decodeBoundary(runtimeEventSchema, {
-		kind: "agent_event",
-		sessionId: input.sessionId,
-		sequence: input.sequence,
-		timestamp:
-			input.event.timestamp ?? input.timestamp ?? new Date().toISOString(),
-		...optionalString("sourceId", input.event.sourceId),
-		...optionalString("runId", input.event.runId),
-		...optionalString("workKey", input.event.workKey),
-		event: input.event.event,
-	});
+}): RuntimeEvent => ({
+	kind: "agent_event",
+	sessionId: input.sessionId,
+	sequence: input.sequence,
+	timestamp:
+		input.event.timestamp ?? input.timestamp ?? new Date().toISOString(),
+	...(input.event.sourceId === undefined
+		? {}
+		: { sourceId: input.event.sourceId }),
+	...(input.event.runId === undefined ? {} : { runId: input.event.runId }),
+	...(input.event.workKey === undefined
+		? {}
+		: { workKey: input.event.workKey }),
+	event: input.event.event,
+});
 
 export const createSessionId = (): string => `session-${randomUUID()}`;
