+++ packages/session/test/agent-runtime.test.ts
@@ -92,74 +92,6 @@ test("runtime publishes appended inner agent events", async () => {
 	await runtime.shutdown();
 });
 
-test("runtime publishes compact inner agent events", async () => {
-	const runtime = makeAgentSessionRuntime({
-		id: "session-1",
-		sources: [],
-		runner,
-	});
-
-	const events = runtime.events();
-	await runtime.appendAgentEvent({
-		sourceId: "runtime-test",
-		runId: "run-1",
-		workKey: "work-1",
-		event: {
-			type: "message_update",
-			assistantMessageEvent: {
-				type: "toolcall_delta",
-				contentIndex: 1,
-				delta: "big body",
-				partial: {
-					content: [
-						{
-							type: "thinking",
-							thinking: "reasoning",
-							thinkingSignature: "encrypted-large-payload",
-						},
-						{
-							type: "toolCall",
-							id: "tool-1",
-							name: "upsert_review_anchor",
-							arguments: { body: "large review body", path: "kept.md" },
-						},
-					],
-				},
-			},
-		},
-	});
-
-	const record = await waitForEvent(
-		events,
-		(entry) => entry.kind === "agent_event",
-	);
-	const encoded = JSON.stringify(record);
-	expect(encoded).not.toContain("encrypted-large-payload");
-	expect(encoded).not.toContain("large review body");
-	expect(record).toMatchObject({
-		kind: "agent_event",
-		event: {
-			type: "message_update",
-			assistantMessageEvent: {
-				type: "toolcall_delta",
-				partial: {
-					content: [
-						{ type: "thinking", thinking: "reasoning" },
-						{
-							type: "toolCall",
-							id: "tool-1",
-							name: "upsert_review_anchor",
-							arguments: { path: "kept.md" },
-						},
-					],
-				},
-			},
-		},
-	});
-
-	await runtime.shutdown();
-});
-
 test("runtime shutdown publishes shutdown and is idempotent", async () => {
 	const runtime = makeAgentSessionRuntime({
 		id: "session-1",
