+++ packages/web/package.json
@@ -14,6 +14,7 @@
 		"@nanostores/query": "^0.3.4",
 		"@nanostores/react": "^1.1.0",
 		"@phosphor-icons/react": "^2.1.10",
+		"@plot/projection": "workspace:*",
 		"@plot/session": "workspace:*",
 		"@tailwindcss/vite": "^4.3.1",
 		"@vitejs/plugin-react": "latest",
