+++ packages/web/test/api.test.ts
@@ -2,7 +2,7 @@ import { expect, test } from "bun:test";
 import {
 	emptyProjection,
 	serializeDashboardProjection,
-} from "@plot/session/projection";
+} from "@plot/projection";
 import { parseProjection } from "../src/data/api.js";
 
 test("parseProjection accepts a real serialized projection envelope", () => {
